==> Notebooks/1. Data Collection(Entrepreneurship-SciTech).ipynb <==
#!/usr/bin/env python
# coding: utf-8
import json
import time
import requests
import pandas as pd    
from requests import get
from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import TimeoutException,StaleElementReferenceException

def add_to_json(data, title, summary, url, category):
    """
    Function for adding data contents to a json file
    """
    data['Article'].append({
        'Title': title,
        'Content': summary,
        'Link': url,
        'Category': category
    })
    return(data['Article'])

def save_json(data, fname):
    """
    Function for saving json file
    """
    with open('{}.json'.format(fname), 'w') as outfile:
        json.dump(data, outfile)

Young Citizen Science portal - Citizen science https://en.wikipedia.org

# MSME micro, small, medium enterprises and entrepreneurship

links = {"https://ryadiybusiness.com/category/%d9%85%d8%b4%d8%a7%d8%b1%d9%8a%d8%b9/page/":7, 
         "https://ryadiybusiness.com/category/%d8%a7%d9%84%d8%aa%d8%b3%d9%88%d9%8a%d9%82-%d9%88%d8%a7%d9%84%d8%a5%d8%a8%d8%aa%d9%83%d8%a7%d8%b1/page":48,
         "https://ryadiybusiness.com/category/%d8%a7%d9%84%d8%ad%d8%a7%d8%b6%d9%86%d8%a7%d8%aa-%d9%88%d8%a7%d9%84%d9%85%d8%b3%d8%b1%d8%b9%d8%a7%d8%aa/page/":44,
         "https://ryadiybusiness.com/category/%d8%a7%d9%84%d8%aa%d9%85%d9%88%d9%8a%d9%84-%d9%88%d8%a7%d9%84%d8%a5%d8%b3%d8%aa%d8%ab%d9%85%d8%a7%d8%b1/page/":89}

categories = ["مشاريع", "التسويق والابتكار", "الحاضنات والمسرعات", "التمويل والاستثمار"]

data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)

l=0
for c in range(len(categories)):
    
    for i in range(1,list(links.values())[c]):
        
        driver.get(list(links.keys())[c]+"{}".format(i))
        soup=BeautifulSoup(driver.page_source,'html.parser')
        stories = soup.findAll("div", {"class": "td-block-span6"})

        for n in range(len(stories)):
            
            title = stories[n].find("h3", {"class":"entry-title td-module-title"}).find('a').text
            url = stories[n].find("h3", {"class":"entry-title td-module-title"}).find('a').attrs['href']
            driver.get(url)
            soup = BeautifulSoup(driver.page_source,'html.parser')
            div = soup.find("div", {"class": "td-post-content tagdiv-type"})
            ps = div.findAll('p')
            content = ""
            for p in ps:
                content += p.text
            category = categories[c]
            
            if content != "" and category != "":
                add_to_json(data, title, content, url, category)
                l += 1
    print(l)

save_json(data, '../Datasets/entrepreneurship/ryadibusiness')

#df = pd.read_csv('../Datasets/ryadibusiness.csv')
df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/entrepreneurship/ryadibusiness.csv", index=False)
df

df['Category'].value_counts()

data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)
driver.get("https://waya.media/arabi/%d8%b4%d8%b1%d9%83%d8%a7%d8%aa-%d9%86%d8%a7%d8%b4%d8%a6%d8%a9/")

while True:
    try:
        showmore=WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.XPATH ,'//*[@id="v/what-are-algorithms-panel"]/div[1]/div/div[6]/div/div[4]/button')))
        showmore.click()
    except TimeoutException:
        break
    except StaleElementReferenceException:
        break
        
for i in range(20):
    driver.execute_script("window.scrollTo(0,document.body.scrollHeight)")
    time.sleep(3)
    
soup=BeautifulSoup(driver.page_source,'html.parser')
media = soup.findAll("div", {"class": "post-media"})

l=0
for n in range(len(media)):
    
    title = media[n].find('img').attrs['alt']
    url = media[n].find('a').attrs['href']
    driver.get(url)
    
    soup = BeautifulSoup(driver.page_source,'html.parser')
    div = soup.find("div", {"class": "entry-content"})
    ps = div.findAll('p')
    content = ""
    for p in ps:
        content += p.text
    if soup.find("span", {"class":"post-category"}) != None:
        category = soup.find("span", {"class":"post-category"}).find("span").text
        if content != "" and category != "":
            add_to_json(data, title, content, url, category)
            l += 1
print(l)
save_json(data, '../Datasets/entrepreneurship/waya')

#df = pd.read_csv('../Datasets/waya.csv')
df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/entrepreneurship/waya.csv", index=False)
df

df['Category'].value_counts()

data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)
driver.get("https://www.youm7.com/Tags/Index?id=149833&tag=%d8%b1%d9%8a%d8%a7%d8%af%d8%a9-%d8%a7%d9%84%d8%a7%d8%b9%d9%85%d8%a7%d9%84")
        
while True:
    try:
        element = driver.find_element(By.CLASS_NAME, 'showMoreNew')
    except:
        print("can't find the 'more' button anymore! Great")
        break;
    else:
        try:
            element.click()
            time.sleep(5)
        except:
            break;
    
soup=BeautifulSoup(driver.page_source,'html.parser')
media = soup.findAll("div", {"class": "col-xs-12 bigOneSec"})

l=0
for n in range(len(media)):
    
    title = media[n].find('h3').find('a').text
    url = "https://www.youm7.com" + media[n].find('h3').find('a').attrs['href']
    driver.get(url)
    
    soup = BeautifulSoup(driver.page_source,'html.parser')
    #title = soup.find("div", {"class": "articlestyled__MastheadContent-rgo6b1-1 hHQxFA"}).find("h1").text
    div = soup.find("div", {"id": "articleBody"})
    ps = div.findAll('p')
    content = ""
    for p in ps:
        content += p.text
    category = "ريادة أعمال"
    if content != "" and category != "":
        add_to_json(data, title, content, url, category)
        l += 1
print(l)
save_json(data, '../Datasets/entrepreneurship/youm7')

df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/entrepreneurship/youm7.csv", index=False)
df

data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)

l=0    
for i in range(1,30):

    driver.get("https://jawlah.co/category/news/page/{}".format(i))
    soup = BeautifulSoup(driver.page_source,'html.parser')
    stories = soup.findAll("h2", {"class": "entry-title"})

    for n in range(len(stories)):

        title = stories[n].find('a').text
        url = stories[n].find('a').attrs['href']
        driver.get(url)
        soup = BeautifulSoup(driver.page_source,'html.parser')
        div = soup.find("div", {"class": "entry-content"})
        ps = div.findAll('p')
        content = ""
        for p in ps:
            content += p.text
        category = "ريادة أعمال"

        if content != "" and category != "":
            add_to_json(data, title, content, url, category)
            l += 1
print(l)

save_json(data, '../Datasets/entrepreneurship/jawlah')

df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/entrepreneurship/jawlah.csv", index=False)
df

data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)
driver.get("https://www.asharqbusiness.com/page/%D8%B1%D9%8A%D8%A7%D8%AF%D8%A9")
        
while True:
    try:
        element = driver.find_element(By.CLASS_NAME, 'btn-load-more')
    except:
        print("can't find the 'more' button anymore! Great")
        break;
    else:
        try:
            element.click()
            time.sleep(5)
        except:
            break;
    
soup=BeautifulSoup(driver.page_source,'html.parser')
stories = soup.findAll("h4", {"class": "heading--post"})

l=0
for n in range(len(stories)):
    
    title = stories[n].find('a').text
    url = "https://www.asharqbusiness.com" + stories[n].find('a').attrs['href']
    driver.get(url)
    
    soup = BeautifulSoup(driver.page_source,'html.parser')
    div = soup.find("div", {"class": "col-12 col-md-9 col-xl-10 article--content"})
    if div != None:
        ps = div.findAll('p')
        content = ""
        for p in ps:
            content += p.text
        category = "ريادة أعمال"
        if content != "":
            add_to_json(data, title, content, url, category)
            l += 1
print(l)
save_json(data, '../Datasets/entrepreneurship/asharqbusiness')

df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/entrepreneurship/asharqbusiness.csv", index=False)
df.head()

# Scroll + Button
data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)
driver.get("https://www.raedaamal.com/category/news/")
  
l=0
while True:
    try:
        element = driver.find_element_by_id('load-more-button')
    except:
        l += 1
        print("can't find the 'more' button anymore! Great")
        if l > 2:
            break;
    else:
        try:
            element.click()
            time.sleep(5)
        except:
            break;
        
  
soup=BeautifulSoup(driver.page_source,'html.parser')
media = soup.findAll("h2", {"class": "entry-title"})

l=0
for n in range(len(media)):
    
    title = media[n].find('a').text
    url = media[n].find('a').attrs['href']
    driver.get(url)
    
    soup = BeautifulSoup(driver.page_source,'html.parser')
    #title = soup.find("div", {"class": "articlestyled__MastheadContent-rgo6b1-1 hHQxFA"}).find("h1").text
    div = soup.find("div", {"class": "s-post-content s-post-small-el bb-mb-el"})
    ps = div.findAll('p')
    content = ""
    for p in ps:
        content += p.text
    category = "ريادة أعمال"
    if content != "":
        add_to_json(data, title, content, url, category)
        l += 1
print(l)
save_json(data, '../Datasets/entrepreneurship/raedaamal')

df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/entrepreneurship/raedaamal.csv", index=False)
df

# check if the added posts are related

data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)

l=0    
for i in range(1,21):

    driver.get("https://www.egyentrepreneur.com/category/entrepreneurship/page/{}/".format(i))
    soup = BeautifulSoup(driver.page_source,'html.parser')
    stories = soup.findAll("h3", {"class": "jeg_post_title"})

    for n in range(len(stories)):

        title = stories[n].find('a').text
        url = stories[n].find('a').attrs['href']
        driver.get(url)
        soup = BeautifulSoup(driver.page_source,'html.parser')
        div = soup.find("div", {"class": "entry-content no-share"})
        if div != None:
            ps = div.findAll('p')
            content = ""
            for p in ps:
                content += p.text
            category = "ريادة أعمال"

            if content != "" and category != "":
                add_to_json(data, title, content, url, category)
                l += 1
print(l)

save_json(data, '../Datasets/entrepreneurship/egyentrepreneur.json')

df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/entrepreneurship/egyentrepreneur.csv", index=False)
df

data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)

l=0    
for i in range(1,13):

    driver.get("https://preneur-masr.com/tag/%d8%b1%d9%8a%d8%a7%d8%af%d8%a9-%d8%a7%d9%84%d8%a3%d8%b9%d9%85%d8%a7%d9%84/page/{}/".format(i))
    soup = BeautifulSoup(driver.page_source,'html.parser')
    stories = soup.findAll("div", {"class": "post-content"})

    for n in range(len(stories)):

        title = stories[n].find('h2').find('a').text
        url = "https://preneur-masr.com" + stories[n].find('h2').find('a').attrs['href']
        driver.get(url)
        
        soup = BeautifulSoup(driver.page_source,'html.parser')
        div = soup.find("div", {"class": "post-content"})
        ps = div.findAll('p')
        content = ""
        for p in ps:
            content += p.text
        category = "ريادة أعمال"

        if content != "":
            add_to_json(data, title, content, url, category)
            l += 1
print(l)

save_json(data, '../Datasets/entrepreneurship/preneur-masr')

df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/entrepreneurship/preneur-masr.csv", index=False)
df

# Science, Technology & innovation

data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)
driver.get("https://arabic.rt.com/technology/news/")

while True:
    try:
        element = driver.find_element_by_class_name('more-button')
    except:
        print("can't find the 'more' button anymore! Great")
        break;
    else:
        driver.execute_script("arguments[0].click();", element)
        time.sleep(5)
   
#element = driver.find_element_by_css('div[class*="loadingWhiteBox"]')
#webdriver.ActionChains(driver).move_to_element(element ).click(element ).perform()
soup=BeautifulSoup(driver.page_source,'html.parser')
stories = soup.findAll("div", {"class": "element"})

l=0
for n in range(len(stories)):
    
    title = stories[n].find('h2').text
    url = 'https://arabic.rt.com/' + stories[n].find('h2').find('a').attrs['href']
    driver.get(url)
    
    soup = BeautifulSoup(driver.page_source,'html.parser')
    div = soup.find("div", {"class": "text js-text js-mediator-article"})
    ps = div.findAll('p')
    content = ""
    for p in ps:
        content += p.text
    category = 'علوم وتكنولوجيا'
    if content != "" and category != "":
        add_to_json(data, title, content, url, category)
        l += 1
print(l)
save_json(data, '../Datasets/Science-Technology/RT-Online')

#df = pd.read_csv('../Datasets/RT-Online.csv')
df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/Science-Technology/RT-Online.csv", index=False)
df

data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)
driver.get("https://asharq.com/category/%D8%AA%D9%83%D9%86%D9%88%D9%84%D9%88%D8%AC%D9%8A%D8%A7/")
        
for i in range(20):
    driver.execute_script("window.scrollTo(0,document.body.scrollHeight)")
    time.sleep(5)
    
soup=BeautifulSoup(driver.page_source,'html.parser')
media = soup.findAll("div", {"class": "gridstyled__GridCardLink-sc-1fz3w0t-20 fLSEMG"})

l=0
for n in range(len(media)):
    
    #title = media[n].find('a').text
    url = "https://asharq.com" + media[n].find('a').attrs['href']
    driver.get(url)
    
    soup = BeautifulSoup(driver.page_source,'html.parser')
    if soup.find("div", {"id": "article_content_body"}) != None:
        title = soup.find("div", {"class": "articlestyled__MastheadContent-rgo6b1-1 hHQxFA"}).find("h1").text
        div = soup.find("div", {"id": "article_content_body"})
        ps = div.findAll('p')
        content = ""
        for p in ps:
            content += p.text
        category = "علوم وتكنولوجيا"
        if content != "" and category != "":
            add_to_json(data, title, content, url, category)
            l += 1
print(l)
save_json(data, '../Datasets/Science-Technology/asharq')

df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/Science-Technology/asharq.csv", index=False)
df

data = {}
data['Article'] = []

service = Service('./chromedriver')
service.start()
driver = webdriver.Remote(service.service_url)
driver.get("https://arabic.sputniknews.com/science/")
        
  
l=0
while True:
    try:
        element = driver.find_element_by_class_name('b-btn.m-more')
    except:
        l += 1
        print("can't find the 'more' button anymore! Great")
        if l > 2:
            break;
    else:
        try:
            element.click()
            time.sleep(5)
        except:
            break;
    
soup=BeautifulSoup(driver.page_source,'html.parser')
media = soup.findAll("div", {"class": "b-stories__title"})

l=0
for n in range(len(media)):
    
    title = media[n].find('h2').find('a').text
    url = "https://arabic.sputniknews.com" + media[n].find('h2').find('a').attrs['href']
    driver.get(url)
    
    soup = BeautifulSoup(driver.page_source,'html.parser')
    if soup.find("div", {"class": "b-article__text"}) != None:
        div = soup.find("div", {"class": "b-article__text"})
        ps = div.findAll('p')
        content = ""
        for p in ps:
            content += p.text
        category = "علوم وتكنولوجيا"
        if content != "" and category != "":
            add_to_json(data, title, content, url, category)
            l += 1
print(l)
save_json(data, '../Datasets/Science-Technology/sputniknews')

df = pd.json_normalize(data['Article'])
df.drop_duplicates(inplace=True)
df.to_csv("../Datasets/Science-Technology/sputniknews.csv", index=False)
df
