==> Notebooks/2. Data Preprocessing.ipynb <==
### Import Libraries

#!/usr/bin/python3

#import nltk
#nltk.download('stopwords')
import warnings
warnings.filterwarnings('ignore')
import os, pickle, re, string
from string import punctuation
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from nltk.corpus import stopwords

### Loading the entrepreneurship datasets

df_entrepreneurship = pd.read_csv("../Datasets/entrepreneurship/asharqbusiness.csv")
df1 = pd.read_csv("../Datasets/entrepreneurship/egyentrepreneur.csv")
df2 = pd.read_csv("../Datasets/entrepreneurship/jawlah.csv")
df3 = pd.read_csv("../Datasets/entrepreneurship/preneur-masr.csv")
df4 = pd.read_csv("../Datasets/entrepreneurship/raedaamal.csv")
df5 = pd.read_csv("../Datasets/entrepreneurship/ryadibusiness.csv") 
df6 = pd.read_csv("../Datasets/entrepreneurship/waya.csv")
df6 = df6[df6['Category'] == 'شركات ناشئة']
df7 = pd.read_csv("../Datasets/entrepreneurship/youm7.csv")
for d in [df1, df2, df3, df4, df5, df6, df7]:
    df_entrepreneurship = df_entrepreneurship.append(d, ignore_index=True)
df_entrepreneurship['Category'] = 'ريادة أعمال'
df_entrepreneurship

### Loading the Science & Technology Data

df_SciTech = pd.read_csv("../Datasets/Science-Technology/RT-Online.csv")
df1 = pd.read_csv("../Datasets/Science-Technology/asharq.csv")
df2 = pd.read_csv("../Datasets/Science-Technology/sputniknews.csv")
for d in [df1, df2]:
    df_SciTech = df_SciTech.append(d, ignore_index=True)
df_SciTech

### Loading the other data

df_other = pd.read_csv("../Datasets/UN_News.csv")
df_other = df_other[df_other['Category'] != 'الثقافة والتعليم']
df_other = df_other[df_other['Category'] != 'التنمية الاقتصادية']
df_other['Category'] = 'أخرى'
df_other

df = pd.DataFrame()
for d in [df_entrepreneurship, df_SciTech, df_other]:
    df = df.append(d.head(2800), ignore_index=True)
df = df.sample(frac=1)
df

# Data Preprocessing

"""
** Remove English and Arabic punctuations
** Normalize different forms of letters using one common form
"""
punctuations = '''`÷×؛<>_()*&^%][ـ،/:"؟.,'{}~¦+|!”…“–ـ''' + string.punctuation

# Arabic stop words with nltk
stop_words = stopwords.words()

arabic_diacritics = re.compile("""
                             ّ    | # Shadda
                             َ    | # Fatha
                             ً    | # Tanwin Fath
                             ُ    | # Damma
                             ٌ    | # Tanwin Damm
                             ِ    | # Kasra
                             ٍ    | # Tanwin Kasr
                             ْ    | # Sukun
                             ـ     # Tatwil/Kashida
                         """, re.VERBOSE)

def preprocess(text):
    '''
    text is an arabic string input
    the preprocessed text is returned
    '''
    #remove punctuations
    translator = str.maketrans('', '', punctuations)
    text = text.translate(translator)
    
    # remove Tashkeel
    text = re.sub(arabic_diacritics, '', text)
    
    #remove longation
    text = re.sub("[إأآا]", "ا", text)
    text = re.sub("ى", "ي", text)
    text = re.sub("ؤ", "ء", text)
    text = re.sub("ئ", "ء", text)
    text = re.sub("ة", "ه", text)
    text = re.sub("گ", "ك", text)
    text = ' '.join(word for word in text.split() if word not in stop_words)

    return text
  
df['Content'] = df['Content'].apply(preprocess)
df['Title'] = df['Title'].apply(preprocess)
df.to_csv('../Datasets/processed_data.csv', index=False)
df.to_json('../Datasets/processed_data.json')

df['Category'].value_counts()

df
