==> Notebooks/4. Predict_new_text.ipynb <==
#!/usr/bin/env python
# coding: utf-8

import pickle
from Preprocess import preprocess

f = open('test.txt', "r")
text = f.read()
model = pickle.load(open('../Artifacts/lr', 'rb'))
Category_lebels = {"أخرى":3 , "علوم وتكنولوجيا":2, "ريادة أعمال":1}
text = preprocess(text)
out = model.predict([text])
print(out)
print("The class of this text is:\t{}".format(list(Category_lebels.keys())[list(Category_lebels.values()).index(out[0])]))
